+++ src/services/fetchers/GoldApiPriceGetter.ts
@@ -13,7 +13,6 @@ import {PriceDataRepository} from '../../repositories/PriceDataRepository.js';
 import TimeService from '../TimeService.js';
 import {GoldApiDataRepository} from '../../repositories/fetchers/GoldApiDataRepository.js';
 import {MappingRepository} from '../../repositories/MappingRepository.js';
-import {FetchersMappingCacheKeys} from './common/FetchersMappingCacheKeys.js';
 
 export interface GoldApiPriceInputParams {
   symbol: string;
@@ -37,12 +36,6 @@ export class GoldApiPriceGetter implements FeedFetcherInterface {
       return {prices: []};
     }
 
-    try {
-      await this.cacheInput(params);
-    } catch (e) {
-      this.logger.error(`${this.logPrefix} failed cache: ${(e as Error).message}`);
-    }
-
     const prices = await this.goldApiDataRepository.getPrices(params, options.timestamp);
 
     // TODO this will be deprecated once we fully switch to DB and have dedicated charts
@@ -55,20 +48,5 @@ export class GoldApiPriceGetter implements FeedFetcherInterface {
     );
 
     return {prices};
-  }
-
-  private async cacheInput(params: GoldApiPriceInputParams[]): Promise<void> {
-    const timestamp = this.timeService.apply();
-    const key = FetchersMappingCacheKeys.GOLD_API_PRICE_PARAMS;
-
-    // const cache = await this.mappingRepository.get(key);
-    // const cachedParams = JSON.parse(cache || '{}');
-    const cachedParams: Record<string, number> = {};
-
-    params.forEach((input) => {
-      cachedParams[`${input.symbol};${input.currency}`.toLowerCase()] = timestamp;
-    });
-
-    await this.mappingRepository.set(key, JSON.stringify(cachedParams));
   }
 }
