+++ src/workers/fetchers/dexes/uniswapV3/UniswapV3FeedsGetter.ts
@@ -1,38 +1,17 @@
 import {inject, injectable} from 'inversify';
-import {Logger} from 'winston';
-
-import TimeService from '../../../../services/TimeService.js';
-import {FeedsType} from '../../../../types/Feed.js';
-import {FeedDataService} from '../../../../services/FeedDataService.js';
 import {ChainsIds} from '../../../../types/ChainsIds.js';
 import {FetcherName} from '../../../../types/fetchers.js';
-import {DeviationLeavesAndFeeds} from '../../../../types/DeviationFeeds.js';
-import {DexProtocolName} from '../../../../types/Dexes.js';
 import {UniswapV3Param} from './interfaces.js';
+import {DeviationFeedsGetter} from '../../_common/DeviationFeedsGetter.js';
 
 @injectable()
 export class UniswapV3FeedsGetter {
-  @inject(FeedDataService) feedDataService!: FeedDataService;
-  @inject(TimeService) timeService!: TimeService;
-  @inject('Logger') logger!: Logger;
+  @inject(DeviationFeedsGetter) feedsGetter!: DeviationFeedsGetter;
 
-  readonly protocol = DexProtocolName.UNISWAP_V3;
   readonly logPrefix = '[UniswapV3FeedsGetter]';
 
   async apply(chainId: ChainsIds): Promise<UniswapV3Param[]> {
-    const feeds = await this.getFeeds(chainId);
-    const uniswapFeeds = this.feedDataService.getParamsByFetcherName<UniswapV3Param>(feeds, FetcherName.UniswapV3);
-    return uniswapFeeds.filter((f) => f.fromChain == chainId);
-  }
-
-  private async getFeeds(chainId: ChainsIds): Promise<DeviationLeavesAndFeeds> {
-    const dataTimestamp = this.timeService.apply();
-    const data = await this.feedDataService.apply(dataTimestamp, FeedsType.DEVIATION_TRIGGER);
-
-    if (data.rejected) {
-      this.logger.info(`${this.logPrefix}[${chainId}] rejected: ${data.rejected}`);
-    }
-
-    return data.feeds as DeviationLeavesAndFeeds;
+    const uniswapFeeds = await this.feedsGetter.apply<UniswapV3Param>(FetcherName.UniswapV3);
+    return uniswapFeeds.filter((f) => f.fromChain == chainId) as UniswapV3Param[];
   }
 }
